+++ page-objects/login/login-page.ts
@@ -1,0 +1,29 @@
+import { Page } from '@playwright/test';
+
+export class LoginPage {
+  private readonly page: Page;
+
+  constructor(page: Page) {
+    this.page = page;
+  }
+
+  async navigateTo(url: string): Promise<void> {
+    await this.page.goto(url);
+  }
+
+  async getTitle(): Promise<string> {
+    return this.page.title();
+  }
+
+  async click(selector: string): Promise<void> {
+    await this.page.click(selector);
+  }
+
+  async type(selector: string, text: string): Promise<void> {
+    await this.page.fill(selector, text);
+  }
+
+  async waitForSelector(selector: string): Promise<void> {
+    await this.page.waitForSelector(selector);
+  }
+}
