+++ page-objects/services/overview/services-overview-page.ts
@@ -1,9 +1,45 @@
-const { expect } = require('@playwright/test');
-const { PaginatedDataTable } = require('../../common/paginated-data-table');
-const { waitForServicesOverviewReady } = require('../../../utils/waits/admin-page-ready');
+import { PaginatedDataTable } from '@page-objects/common/paginated-data-table';
+import { Locator, Page, expect } from '@playwright/test';
 
-class ServicesOverviewPage {
-  constructor(page) {
+import { RouteReadyOptions } from '@utils/interfaces';
+import { waitForServicesOverviewReady } from '@utils/waits/admin-page-ready';
+
+interface FindServiceRowOptions {
+  readonly pageSize?: string;
+  readonly maxPages?: number;
+}
+
+interface ExpectServiceRowOptions {
+  readonly pageSize?: string;
+  readonly timeout?: number;
+}
+
+export class ServicesOverviewPage {
+  private readonly page: Page;
+
+  private readonly headingServices: Locator;
+  private readonly headingGeneralServices: Locator;
+
+  private readonly buttonImportMany: Locator;
+  private readonly buttonExportMany: Locator;
+  private readonly buttonCreateNewService: Locator;
+
+  private readonly buttonConfirmDelete: Locator;
+  private readonly buttonCancelDelete: Locator;
+
+  private readonly tableServices: Locator;
+  private readonly tableGeneralServices: Locator;
+
+  private readonly thName: Locator;
+  private readonly thDescription: Locator;
+  private readonly thStatus: Locator;
+
+  private readonly selectPageSizeServices: Locator;
+  private readonly selectPageSizeGeneralServices: Locator;
+
+  private readonly servicesTable: PaginatedDataTable;
+
+  constructor(page: Page) {
     this.page = page;
 
     this.headingServices = this.page.getByRole('heading', { name: 'Teenused', exact: true });
@@ -14,17 +50,36 @@ class ServicesOverviewPage {
     this.buttonCreateNewService = this.page.getByRole('button', { name: 'Loo uus teenus', exact: true });
 
     this.buttonConfirmDelete = this.page.getByRole('dialog').getByRole('button', { name: 'Kustuta' }).last();
-    this.buttonCancelDelete = this.page.getByRole('dialog').getByRole('button', { name: /tühista|cancel/i }).first();
+    this.buttonCancelDelete = this.page
+      .getByRole('dialog')
+      .getByRole('button', { name: /tühista|cancel/i })
+      .first();
 
-    this.tableServices = this.page.getByTestId('services-table').or(this.page.locator('table.data-table').nth(0)).or(this.page.locator('table').nth(0)).first();
-    this.tableGeneralServices = this.page.getByTestId('general-services-table').or(this.page.locator('table.data-table').nth(1)).or(this.page.locator('table').nth(1)).first();
+    this.tableServices = this.page
+      .getByTestId('services-table')
+      .or(this.page.locator('table.data-table').nth(0))
+      .or(this.page.locator('table').nth(0))
+      .first();
+    this.tableGeneralServices = this.page
+      .getByTestId('general-services-table')
+      .or(this.page.locator('table.data-table').nth(1))
+      .or(this.page.locator('table').nth(1))
+      .first();
 
     this.thName = this.page.getByRole('columnheader', { name: 'Nimetus' });
     this.thDescription = this.page.getByRole('columnheader', { name: 'Kirjeldus' });
     this.thStatus = this.page.getByRole('columnheader', { name: 'Olek' });
 
-    this.selectPageSizeServices = this.page.getByTestId('services-page-size').or(this.page.getByRole('combobox', { name: /Kuvan korraga/i }).nth(0)).or(this.page.locator('select').nth(0)).first();
-    this.selectPageSizeGeneralServices = this.page.getByTestId('general-services-page-size').or(this.page.getByRole('combobox', { name: /Kuvan korraga/i }).nth(1)).or(this.page.locator('select').nth(1)).first();
+    this.selectPageSizeServices = this.page
+      .getByTestId('services-page-size')
+      .or(this.page.getByRole('combobox', { name: /Kuvan korraga/i }).nth(0))
+      .or(this.page.locator('select').nth(0))
+      .first();
+    this.selectPageSizeGeneralServices = this.page
+      .getByTestId('general-services-page-size')
+      .or(this.page.getByRole('combobox', { name: /Kuvan korraga/i }).nth(1))
+      .or(this.page.locator('select').nth(1))
+      .first();
 
     this.servicesTable = new PaginatedDataTable(this.page, {
       table: this.tableServices,
@@ -34,58 +89,58 @@ class ServicesOverviewPage {
     });
   }
 
-  async waitForReady(options = {}) {
+  async waitForReady(options: RouteReadyOptions = {}): Promise<void> {
     await waitForServicesOverviewReady(this.page, options);
     await this.servicesTable.waitUntilReady(options);
   }
 
-  getServiceRow(serviceTitle) {
+  getServiceRow(serviceTitle: string): Locator {
     return this.servicesTable.getRowByText(serviceTitle);
   }
 
-  async findServiceRow(serviceTitle, options = {}) {
+  async findServiceRow(serviceTitle: string, options: FindServiceRowOptions = {}): Promise<Locator> {
     await this.waitForReady();
     await this.servicesTable.ensureRowsPerPage(options.pageSize);
     return this.servicesTable.findRowAcrossPages(serviceTitle, options);
   }
 
-  getFirstTableRow(table = this.tableServices) {
+  getFirstTableRow(table: Locator = this.tableServices): Locator {
     return table.locator('tbody').locator('tr').first();
   }
 
-  getRowColumns(rowOrTitle) {
-    const isLocator = rowOrTitle && typeof rowOrTitle.locator === 'function';
+  getRowColumns(rowOrTitle: Locator | string | number): Locator {
+    const isLocator = typeof rowOrTitle === 'object' && typeof rowOrTitle.locator === 'function';
     const isTextLike = typeof rowOrTitle === 'string' || typeof rowOrTitle === 'number';
-    const row = isLocator ? rowOrTitle : (isTextLike ? this.getServiceRow(String(rowOrTitle)) : null);
+    const row = isLocator ? rowOrTitle : isTextLike ? this.getServiceRow(String(rowOrTitle)) : null;
     if (!row || typeof row.locator !== 'function') {
       throw new Error(`Expected a row locator or service title, received: ${typeof rowOrTitle}`);
     }
     return row.locator('td');
   }
 
-  async setServicesRowsPerPageTo50() {
+  async setServicesRowsPerPageTo50(): Promise<void> {
     await this.waitForReady();
     await this.servicesTable.ensureRowsPerPage('50');
   }
 
-  async assertServiceRowVisible(serviceTitle, options = {}) {
+  async assertServiceRowVisible(serviceTitle: string, options: ExpectServiceRowOptions = {}): Promise<void> {
     await this.waitForReady();
     await this.servicesTable.expectRowVisible(serviceTitle, options);
   }
 
-  async assertRowDeleted(serviceTitle) {
+  async assertRowDeleted(serviceTitle: string): Promise<void> {
     await this.waitForReady();
     await this.servicesTable.expectRowDeleted(serviceTitle);
   }
 
-  async clickCreateNew() {
+  async clickCreateNew(): Promise<void> {
     await this.waitForReady();
     await this.buttonCreateNewService.click();
     await this.page.waitForURL(/services\/newService/i, { timeout: 15000 }).catch(() => {});
     await this.page.waitForLoadState('domcontentloaded').catch(() => {});
   }
 
-  async clickEdit(serviceTitle) {
+  async clickEdit(serviceTitle: string): Promise<void> {
     const row = await this.findServiceRow(serviceTitle);
     const button = this.servicesTable.getActionButton(row, 'Muuda');
     await expect(button).toBeVisible();
@@ -94,21 +149,22 @@ class ServicesOverviewPage {
     await this.page.waitForLoadState('domcontentloaded').catch(() => {});
   }
 
-  async clickExport(serviceTitle) {
+  async clickExport(serviceTitle: string): Promise<void> {
     const row = await this.findServiceRow(serviceTitle);
     const button = this.servicesTable.getActionButton(row, 'Ekspordi');
     await expect(button).toBeVisible();
     await button.click();
   }
 
-  async deleteService(serviceTitle) {
+  async deleteService(serviceTitle: string): Promise<void> {
     const row = await this.findServiceRow(serviceTitle);
     await expect(row.first()).toBeVisible({ timeout: 10000 });
 
     const targetRow = row.first();
     const overviewDeleteButton = this.servicesTable.getActionButton(targetRow, 'Kustuta').first();
-    const canDeleteFromOverview = await overviewDeleteButton.isVisible().catch(() => false)
-      && await overviewDeleteButton.isEnabled().catch(() => false);
+    const canDeleteFromOverview =
+      (await overviewDeleteButton.isVisible().catch(() => false)) &&
+      (await overviewDeleteButton.isEnabled().catch(() => false));
 
     if (canDeleteFromOverview) {
       await overviewDeleteButton.scrollIntoViewIfNeeded().catch(() => {});
@@ -126,12 +182,16 @@ class ServicesOverviewPage {
       await headerDeleteButton.click({ force: true }).catch(() => {});
     }
 
-    const deleteDialogVisible = await this.page.getByRole('dialog').isVisible().catch(() => false);
+    const deleteDialogVisible = await this.page
+      .getByRole('dialog')
+      .isVisible()
+      .catch(() => false);
     if (deleteDialogVisible) {
       await this.buttonConfirmDelete.click({ force: true }).catch(() => {});
     }
 
-    const deletedFromCurrentView = await expect(this.getServiceRow(serviceTitle)).toHaveCount(0, { timeout: 10000 })
+    const deletedFromCurrentView = await expect(this.getServiceRow(serviceTitle))
+      .toHaveCount(0, { timeout: 10000 })
       .then(() => true)
       .catch(() => false);
     if (!deletedFromCurrentView) {
@@ -142,12 +202,15 @@ class ServicesOverviewPage {
     }
   }
 
-  async hasServiceRow(serviceTitle, { pageSize = this.servicesTable.defaultPageSize } = {}) {
+  async hasServiceRow(
+    serviceTitle: string,
+    { pageSize = this.servicesTable.defaultPageSize }: { pageSize?: string } = {},
+  ): Promise<boolean> {
     const row = await this.findServiceRow(serviceTitle, { pageSize }).catch(() => null);
-    return row ? await row.count().catch(() => 0) > 0 : false;
+    return row ? (await row.count().catch(() => 0)) > 0 : false;
   }
 
-  async deleteServiceIfExists(serviceTitle) {
+  async deleteServiceIfExists(serviceTitle: string): Promise<boolean> {
     try {
       if (!(await this.hasServiceRow(serviceTitle))) {
         return false;
@@ -163,50 +226,56 @@ class ServicesOverviewPage {
       if (!stillExists) {
         return true;
       }
-      console.warn(`Cleanup failed for service "${serviceTitle}": ${error?.message || error}`);
+      console.warn(`Cleanup failed for service "${serviceTitle}": ${error instanceof Error ? error.message : error}`);
       return false;
     }
   }
 
-  async assertServiceNameExists() {
+  async assertServiceNameExists(): Promise<void> {
     await this.waitForReady();
     const firstCell = this.getFirstTableRow().locator('td').first();
     await expect(firstCell).toBeVisible();
     await expect(firstCell).toContainText(/\S/);
   }
 
-  async assertDescriptionFieldExists() {
+  async assertDescriptionFieldExists(): Promise<void> {
     await this.waitForReady();
     await expect(this.getRowColumns(this.getFirstTableRow()).nth(1)).toBeVisible();
   }
 
-  async assertStatusExists() {
+  async assertStatusExists(): Promise<void> {
     await this.waitForReady();
     const statuses = ['Mustand', 'Valmis', 'Aktiivne'];
     await expect(this.getRowColumns(this.getFirstTableRow()).nth(2)).toContainText(new RegExp(statuses.join('|')));
   }
 
-  async assertStatusReady(rowOrTitle) {
+  async assertStatusReady(rowOrTitle: Locator | string | number): Promise<void> {
     await this.waitForReady();
     await expect(this.getRowColumns(rowOrTitle).nth(2)).toContainText('Valmis');
   }
 
-  async assertEditButtonExists() {
+  async assertEditButtonExists(): Promise<void> {
     await this.waitForReady();
-    await expect(this.getRowColumns(this.getFirstTableRow()).nth(3).getByRole('button', { name: 'Muuda' })).toBeVisible();
+    await expect(
+      this.getRowColumns(this.getFirstTableRow()).nth(3).getByRole('button', { name: 'Muuda' }),
+    ).toBeVisible();
   }
 
-  async assertExportButtonExists() {
+  async assertExportButtonExists(): Promise<void> {
     await this.waitForReady();
-    await expect(this.getRowColumns(this.getFirstTableRow()).nth(4).getByRole('button', { name: 'Ekspordi' })).toBeVisible();
+    await expect(
+      this.getRowColumns(this.getFirstTableRow()).nth(4).getByRole('button', { name: 'Ekspordi' }),
+    ).toBeVisible();
   }
 
-  async assertDeleteButtonExists() {
+  async assertDeleteButtonExists(): Promise<void> {
     await this.waitForReady();
-    await expect(this.getRowColumns(this.getFirstTableRow()).nth(5).getByRole('button', { name: 'Kustuta' })).toBeVisible();
+    await expect(
+      this.getRowColumns(this.getFirstTableRow()).nth(5).getByRole('button', { name: 'Kustuta' }),
+    ).toBeVisible();
   }
 
-  async assertPageSizeVisibleServices() {
+  async assertPageSizeVisibleServices(): Promise<void> {
     await this.waitForReady();
     if (await this.selectPageSizeServices.isVisible().catch(() => false)) {
       await expect(this.selectPageSizeServices).toBeVisible();
@@ -215,7 +284,7 @@ class ServicesOverviewPage {
     await expect(this.tableServices).toBeVisible();
   }
 
-  async assertPageSizeVisibleGeneralServices() {
+  async assertPageSizeVisibleGeneralServices(): Promise<void> {
     await this.waitForReady();
     if (await this.selectPageSizeGeneralServices.isVisible().catch(() => false)) {
       await expect(this.selectPageSizeGeneralServices).toBeVisible();
@@ -224,5 +293,3 @@ class ServicesOverviewPage {
     await expect(this.tableGeneralServices).toBeVisible();
   }
 }
-
-module.exports = { ServicesOverviewPage };
